admin - category

+++ src/main/java/com/eastflag/nnc/board/model/Board.java
@@ -1,5 +1,6 @@
 package com.eastflag.nnc.board.model;
 
+import com.eastflag.nnc.board_category.model.BoardCategory;
 import com.eastflag.nnc.common.BaseEntity;
 import com.eastflag.nnc.user.model.User;
 import jakarta.persistence.*;
@@ -15,7 +16,7 @@ import lombok.*;
 public class Board extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(length = 100, nullable = false)
     private String title;
